Create VPRM_month.py
This is the main script which is based on [Mahadevan, P., S. C. Wofsy, D. M. Matross, X. Xiao, A. L. Dunn, J. C. Lin, C. Gerbig, J. W. Munger, V. Y. Chow,and E. W. Gottlieb (2008), A satellite-based biosphere parameterization for net ecosystem CO2exchange: Vegetation Photosynthesis and Respiration Model (VPRM), Global Biogeochem. Cycles, 22, GB2005, doi:10.1029/2006GB002735.] and provides calculation of CO2 biospheric fluxes. In the script we use specific a priori data (surface reflectance, air temperature, photosynthetically active radiation and some constant values which depending on various conditions) which are provided with uniform spatial resolution so it can be simply used for the VPRM calculation on 2D model grid. The bash-scripts which are implemented to obtain the prior data are also provided in this repository. The prior data which we are using in this script allow us to calculate the CO2 biospheric fluxes with 0.25 deg spatial resolution every 3 h.